Migrate builtin functions from evaluator to object module
So that builtins are not directly tied to the evaluator and can be used
in the VM module without it depending on the evaluator.

Have true false and null objects defined in object module so only one
instance of each is required.

builtins are now stored as a slice rather than as a map. This is to
ensure consistent ordering for use in referencing builtins from the
VM.