Finishing touches
 - Update README
 - Fix small bugs
 - Clean up main file
 - Fix major bug missed in the original implementation

The major bug in question was that function calls would break if one of
their arguments was an anonymous closure that captured variables from
its enclosing scope.

The cause was that the original implementation adds 'free variables' to
the stack before executing the OpClosure instruction, which inside a
function call would add more objects onto the stack then expected. When
executing the OpCall instruction, it would expect n number of arguments,
but the stack would have n+k values where k is the number of free
variables captured by the closure. The result was that the OpCall
instruction would be looking at the wrong item on the stack to find the
closure object it was trying to call.

The Fix for this was to simply pop the free variables off the stack
after they'd been added to the closure during its creation.